+++ src/sceneload.cpp
@@ -29,56 +29,56 @@ bool load_scene(const std::string& filepath) {
         
         //Read in a camera
         if (strcmp(lineHeader, "camera") == 0) {
-            glm::vec3 temp_pos(-1.0f);
+            glm::dvec3 temp_pos(-1.0f);
             int temp_fov = 0;
             int temp_f = 0;
-            float temp_a = 0.0;
+            double temp_a = 0.0;
             for (int i = 0; i < 4; i++) {
                 fscanf(file, "%s: ", lineHeader);
                 if (strcmp(lineHeader, "pos:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_pos.x, &temp_pos.y, &temp_pos.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_pos.x, &temp_pos.y, &temp_pos.z);
                 } else if (strcmp(lineHeader, "fov:") == 0) {
                     fscanf(file, "%i\n", &temp_fov);
                 } else if (strcmp(lineHeader, "f:") == 0) {
                     fscanf(file, "%i\n", &temp_f);
                 } else if (strcmp(lineHeader, "a:") == 0) {
-                    fscanf(file, "%f\n", &temp_a);
+                    fscanf(file, "%lf\n", &temp_a);
                 } else {
                     std::cout << lineHeader << "\nCamera is missing info!" << std::endl;
                     return false;
                 }
             }
-            printf("Camera Specs: \n  Position: %s\n  FoV: %i\n  Aspect Ratio: %f\n  Focal Length: %i\n", glm::to_string(temp_pos).c_str(), temp_fov, temp_a, temp_f);
+            printf("Camera Specs: \n  Position: %s\n  FoV: %i\n  Aspect Ratio: %lf\n  Focal Length: %i\n", glm::to_string(temp_pos).c_str(), temp_fov, temp_a, temp_f);
         }
         
         //Read in the plane
         else if (strcmp(lineHeader, "plane") == 0) {
-            glm::vec3 temp_norm;
-            glm::vec3 temp_pos;
-            glm::vec3 temp_amb;
-            glm::vec3 temp_dif;
-            glm::vec3 temp_spe;
-            float temp_shi = 0.0;
+            glm::dvec3 temp_norm;
+            glm::dvec3 temp_pos;
+            glm::dvec3 temp_amb;
+            glm::dvec3 temp_dif;
+            glm::dvec3 temp_spe;
+            double temp_shi = 0.0;
             for (int i = 0; i < 6; i++) {
                 fscanf(file, "%s: ", lineHeader);
                 if (strcmp(lineHeader, "nor:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_norm.x, &temp_norm.y, &temp_norm.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_norm.x, &temp_norm.y, &temp_norm.z);
                 } else if (strcmp(lineHeader, "pos:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_pos.x, &temp_pos.y, &temp_pos.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_pos.x, &temp_pos.y, &temp_pos.z);
                 } else if (strcmp(lineHeader, "amb:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_amb.x, &temp_amb.y, &temp_amb.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_amb.x, &temp_amb.y, &temp_amb.z);
                 } else if (strcmp(lineHeader, "dif:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_dif.x, &temp_dif.y, &temp_dif.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_dif.x, &temp_dif.y, &temp_dif.z);
                 } else if (strcmp(lineHeader, "spe:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_spe.x, &temp_spe.y, &temp_spe.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_spe.x, &temp_spe.y, &temp_spe.z);
                 } else if (strcmp(lineHeader, "shi:") == 0) {
-                    fscanf(file, "%f\n", &temp_shi);
+                    fscanf(file, "%lf\n", &temp_shi);
                 } else {
                     std::cout << "Plane is missing info" << std::endl;
                     return false;
                 }
             }
-            printf("Plane Specs: \n  Norm: %s\n  Position: %s\n  Ambient: %s\n  Diffuse: %s\n  Specular: %s\n  Shine: %f\n",
+            printf("Plane Specs: \n  Norm: %s\n  Position: %s\n  Ambient: %s\n  Diffuse: %s\n  Specular: %s\n  Shine: %lf\n",
                    glm::to_string(temp_norm).c_str(),
                    glm::to_string(temp_pos).c_str(),
                    glm::to_string(temp_amb).c_str(),
@@ -89,32 +89,32 @@ bool load_scene(const std::string& filepath) {
         
         //Read in a sphere
         else if (strcmp(lineHeader, "sphere") == 0) {
-            glm::vec3 temp_pos;
+            glm::dvec3 temp_pos;
             int temp_rad = 0;
-            glm::vec3 temp_amb;
-            glm::vec3 temp_dif;
-            glm::vec3 temp_spe;
-            float temp_shi = 0.0;
+            glm::dvec3 temp_amb;
+            glm::dvec3 temp_dif;
+            glm::dvec3 temp_spe;
+            double temp_shi = 0.0;
             for (int i = 0; i < 6; i++) {
                 fscanf(file, "%s: ", lineHeader);
                 if (strcmp(lineHeader, "pos:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_pos.x, &temp_pos.y, &temp_pos.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_pos.x, &temp_pos.y, &temp_pos.z);
                 } else if (strcmp(lineHeader, "rad:") == 0) {
                     fscanf(file, "%i\n", &temp_rad);
                 } else if (strcmp(lineHeader, "amb:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_amb.x, &temp_amb.y, &temp_amb.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_amb.x, &temp_amb.y, &temp_amb.z);
                 } else if (strcmp(lineHeader, "dif:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_dif.x, &temp_dif.y, &temp_dif.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_dif.x, &temp_dif.y, &temp_dif.z);
                 } else if (strcmp(lineHeader, "spe:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_spe.x, &temp_spe.y, &temp_spe.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_spe.x, &temp_spe.y, &temp_spe.z);
                 } else if (strcmp(lineHeader, "shi:") == 0) {
-                    fscanf(file, "%f\n", &temp_shi);
+                    fscanf(file, "%lf\n", &temp_shi);
                 } else {
                     std::cout << "Sphere is missing info" << std::endl;
                     return false;
                 }
             }
-            printf("Sphere Specs: \n  Position: %s\n  Rad: %i\n  Ambient:%s\n  Diffuse: %s\n  Specular: %s\n  Shine: %f\n",
+            printf("Sphere Specs: \n  Position: %s\n  Rad: %i\n  Ambient:%s\n  Diffuse: %s\n  Specular: %s\n  Shine: %lf\n",
                    glm::to_string(temp_pos).c_str(),
                    temp_rad,
                    glm::to_string(temp_amb).c_str(),
@@ -125,14 +125,14 @@ bool load_scene(const std::string& filepath) {
         
         //Read in light
         else if (strcmp(lineHeader, "light") == 0) {
-            glm::vec3 temp_pos;
-            glm::vec3 temp_col;
+            glm::dvec3 temp_pos;
+            glm::dvec3 temp_col;
             for (int i = 0; i < 2; i++) {
                 fscanf(file, "%s: ", lineHeader);
                 if (strcmp(lineHeader, "pos:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_pos.x, &temp_pos.y, &temp_pos.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_pos.x, &temp_pos.y, &temp_pos.z);
                 } else if (strcmp(lineHeader, "col:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_col.x, &temp_col.y, &temp_col.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_col.x, &temp_col.y, &temp_col.z);
                 } else {
                     std::cout << "Light is missing info" << std::endl;
                     return false;
@@ -179,20 +179,20 @@ bool load_scene(const std::string& filepath, std::vector<Object*> & things, std:
         
         //Read in a camera
         if (strcmp(lineHeader, "camera") == 0) {
-            glm::vec3 temp_pos(-1.0f);
+            glm::dvec3 temp_pos(-1.0f);
             int temp_fov = 0;
             int temp_f = 0;
-            float temp_a = 0.0;
+            double temp_a = 0.0;
             for (int i = 0; i < 4; i++) {
                 fscanf(file, "%s: ", lineHeader);
                 if (strcmp(lineHeader, "pos:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_pos.x, &temp_pos.y, &temp_pos.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_pos.x, &temp_pos.y, &temp_pos.z);
                 } else if (strcmp(lineHeader, "fov:") == 0) {
                     fscanf(file, "%i\n", &temp_fov);
                 } else if (strcmp(lineHeader, "f:") == 0) {
                     fscanf(file, "%i\n", &temp_f);
                 } else if (strcmp(lineHeader, "a:") == 0) {
-                    fscanf(file, "%f\n", &temp_a);
+                    fscanf(file, "%lf\n", &temp_a);
                 } else {
                     std::cout << lineHeader << "\nCamera is missing info!" << std::endl;
                     return false;
@@ -203,26 +203,26 @@ bool load_scene(const std::string& filepath, std::vector<Object*> & things, std:
         
         //Read in the plane
         else if (strcmp(lineHeader, "plane") == 0) {
-            glm::vec3 temp_norm;
-            glm::vec3 temp_pos;
-            glm::vec3 temp_amb;
-            glm::vec3 temp_dif;
-            glm::vec3 temp_spe;
-            float temp_shi = 0.0;
+            glm::dvec3 temp_norm;
+            glm::dvec3 temp_pos;
+            glm::dvec3 temp_amb;
+            glm::dvec3 temp_dif;
+            glm::dvec3 temp_spe;
+            double temp_shi = 0.0;
             for (int i = 0; i < 6; i++) {
                 fscanf(file, "%s: ", lineHeader);
                 if (strcmp(lineHeader, "nor:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_norm.x, &temp_norm.y, &temp_norm.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_norm.x, &temp_norm.y, &temp_norm.z);
                 } else if (strcmp(lineHeader, "pos:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_pos.x, &temp_pos.y, &temp_pos.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_pos.x, &temp_pos.y, &temp_pos.z);
                 } else if (strcmp(lineHeader, "amb:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_amb.x, &temp_amb.y, &temp_amb.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_amb.x, &temp_amb.y, &temp_amb.z);
                 } else if (strcmp(lineHeader, "dif:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_dif.x, &temp_dif.y, &temp_dif.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_dif.x, &temp_dif.y, &temp_dif.z);
                 } else if (strcmp(lineHeader, "spe:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_spe.x, &temp_spe.y, &temp_spe.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_spe.x, &temp_spe.y, &temp_spe.z);
                 } else if (strcmp(lineHeader, "shi:") == 0) {
-                    fscanf(file, "%f\n", &temp_shi);
+                    fscanf(file, "%lf\n", &temp_shi);
                 } else {
                     std::cout << "Plane is missing info" << std::endl;
                     return false;
@@ -234,26 +234,26 @@ bool load_scene(const std::string& filepath, std::vector<Object*> & things, std:
         
         //Read in a sphere
         else if (strcmp(lineHeader, "sphere") == 0) {
-            glm::vec3 temp_pos;
+            glm::dvec3 temp_pos;
             int temp_rad = 0;
-            glm::vec3 temp_amb;
-            glm::vec3 temp_dif;
-            glm::vec3 temp_spe;
-            float temp_shi = 0.0;
+            glm::dvec3 temp_amb;
+            glm::dvec3 temp_dif;
+            glm::dvec3 temp_spe;
+            double temp_shi = 0.0;
             for (int i = 0; i < 6; i++) {
                 fscanf(file, "%s: ", lineHeader);
                 if (strcmp(lineHeader, "pos:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_pos.x, &temp_pos.y, &temp_pos.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_pos.x, &temp_pos.y, &temp_pos.z);
                 } else if (strcmp(lineHeader, "rad:") == 0) {
                     fscanf(file, "%i\n", &temp_rad);
                 } else if (strcmp(lineHeader, "amb:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_amb.x, &temp_amb.y, &temp_amb.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_amb.x, &temp_amb.y, &temp_amb.z);
                 } else if (strcmp(lineHeader, "dif:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_dif.x, &temp_dif.y, &temp_dif.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_dif.x, &temp_dif.y, &temp_dif.z);
                 } else if (strcmp(lineHeader, "spe:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_spe.x, &temp_spe.y, &temp_spe.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_spe.x, &temp_spe.y, &temp_spe.z);
                 } else if (strcmp(lineHeader, "shi:") == 0) {
-                    fscanf(file, "%f\n", &temp_shi);
+                    fscanf(file, "%lf\n", &temp_shi);
                 } else {
                     std::cout << "Sphere is missing info" << std::endl;
                     return false;
@@ -265,14 +265,14 @@ bool load_scene(const std::string& filepath, std::vector<Object*> & things, std:
         
         //Read in light
         else if (strcmp(lineHeader, "light") == 0) {
-            glm::vec3 temp_pos;
-            glm::vec3 temp_col;
+            glm::dvec3 temp_pos;
+            glm::dvec3 temp_col;
             for (int i = 0; i < 2; i++) {
                 fscanf(file, "%s: ", lineHeader);
                 if (strcmp(lineHeader, "pos:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_pos.x, &temp_pos.y, &temp_pos.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_pos.x, &temp_pos.y, &temp_pos.z);
                 } else if (strcmp(lineHeader, "col:") == 0) {
-                    fscanf(file, "%f %f %f\n", &temp_col.x, &temp_col.y, &temp_col.z);
+                    fscanf(file, "%lf %lf %lf\n", &temp_col.x, &temp_col.y, &temp_col.z);
                 } else {
                     std::cout << "Light is missing info" << std::endl;
                     return false;
