+++ src/util.cpp
@@ -18,9 +18,9 @@ const char* get_name(std::string name, std::string ext) {
 }
 
 //Draw a 150x150 pink square in the centre of an image
-void draw_square(cimg_library::CImg<float> &image, int width, int height) {
+void draw_square(cimg_library::CImg<double> &image, int width, int height) {
     //Draw a pink square in the centre
-    glm::vec3 color(64, 0, 77);
+    glm::dvec3 color(64, 0, 77);
     for (int j = (width/2)-75; j < (width/2)+75; j++) {
         for (int k = (height/2)-75; k < (height/2)+75; k++) {
             image(j, k, 0) = color.x;
@@ -31,7 +31,7 @@ void draw_square(cimg_library::CImg<float> &image, int width, int height) {
 }
 
 //Set the colour of a given pixel
-void draw(cimg_library::CImg<float> &image, int w, int h, glm::vec3 color) {
+void draw(cimg_library::CImg<double> &image, int w, int h, glm::dvec3 color) {
     //Color in a given pixel
     image(w, h, 0) = color.r*255;
     image(w, h, 1) = color.g*255;
@@ -39,7 +39,7 @@ void draw(cimg_library::CImg<float> &image, int w, int h, glm::vec3 color) {
 }
 
 //clips a vec3
-glm::vec3 clip(glm::vec3 vec, float lo, float hi) {
+glm::dvec3 clip(glm::dvec3 vec, double lo, double hi) {
     vec.x = (vec.x > hi) ? hi : vec.x;
     vec.y = (vec.y > hi) ? hi : vec.y;
     vec.z = (vec.z > hi) ? hi : vec.z;
@@ -54,7 +54,7 @@ glm::vec3 clip(glm::vec3 vec, float lo, float hi) {
 
 
 //returns true if p1 is closer to p0
-bool is_closer(const glm::vec3 p0, const glm::vec3 p1, const glm::vec3 p2) {
+bool is_closer(const glm::dvec3 p0, const glm::dvec3 p1, const glm::dvec3 p2) {
     if ((glm::distance(p0, p1) > glm::distance(p0, p2))) {
         return false;
     }
@@ -64,23 +64,23 @@ bool is_closer(const glm::vec3 p0, const glm::vec3 p1, const glm::vec3 p2) {
 
 
 //general colour calculation function. here to save time during tweaking.
-glm::vec3 calculate_colour(glm::vec3 lightPos, glm::vec3 lightCol, glm::vec3 camPos, glm::vec3 p0, glm::vec3 norm, glm::vec3 diffIn, glm::vec3 specIn, glm::vec3 ambIn, float shineIn) {
+glm::dvec3 calculate_colour(glm::dvec3 lightPos, glm::dvec3 lightCol, glm::dvec3 camPos, glm::dvec3 p0, glm::dvec3 norm, glm::dvec3 diffIn, glm::dvec3 specIn, glm::dvec3 ambIn, double shineIn) {
     
     //Get missing directions and reflections
-    glm::vec3 view_dir = glm::normalize(camPos - p0);
-    glm::vec3 light_dir = glm::normalize(lightPos - p0);
-    glm::vec3 reflection = glm::reflect(norm, view_dir);
+    glm::dvec3 view_dir = glm::normalize(camPos - p0);
+    glm::dvec3 light_dir = glm::normalize(lightPos - p0);
+    glm::dvec3 reflection = glm::reflect(norm, view_dir);
     
     //Get alpha and theta
-    float alpha = glm::dot(view_dir, reflection);
-    float theta = glm::dot(norm, light_dir);
+    double alpha = glm::dot(view_dir, reflection);
+    double theta = glm::dot(norm, light_dir);
     
     //actually calculate specular and diffuse
-    glm::vec3 spec_fin = glm::pow(glm::max(alpha, 0.0f), shineIn) * specIn;
-    glm::vec3 diff_fin = glm::max(theta, 0.0f) * diffIn;
+    glm::dvec3 spec_fin = glm::pow(glm::max(alpha, 0.0), shineIn) * specIn;
+    glm::dvec3 diff_fin = glm::max(theta, 0.0) * diffIn;
     
     //finally calculate and return the colour
-    glm::vec3 colour = (ambIn + diff_fin + spec_fin) * lightCol;
+    glm::dvec3 colour = (ambIn + diff_fin + spec_fin) * lightCol;
     colour = clip(colour, 0.0f, 1.0f);
     return colour;
 };
